bug fix: use the same prefix for build.cypher
The CYPHER_EXECUTION step attempts to read the build.cypher in the
snomed_g_bin directory, however it was created in the output
directory by the previous (TEMPLATE_PROCESSING) step.

Therefore, the directory prefix in CYPHER_EXECUTION was made to
match the one in TEMPLATE_PROCESSING: removed %s and snomed_g_bin.