Replaced mysql-loader-with-optimized-views with completely new version
Old version of mysql-loader-with-optimized-views replaced with complete new version. The new version contains many enhancement including easy to use scripts for MacOS and Windows that rather necessary data from user, inserts these parameters in a temporary copy of the loader script and runs this script.
The new version was developed on a separate GitHub branch in the SnomedRfsMySql folder. Detailed instruction on use are at http://snomed.org/snomedmysql